remove: PhaseData class
